[SWM-300] Refactor : move review model to mapper package
