+++ src/main/java/com/m9d/sroom/lecture/service/LectureServiceV2.java
@@ -3,10 +3,10 @@ package com.m9d.sroom.lecture.service;
 import com.m9d.sroom.course.exception.CourseNotMatchException;
 import com.m9d.sroom.course.exception.CourseVideoNotFoundException;
 import com.m9d.sroom.course.repository.CourseRepository;
-import com.m9d.sroom.global.model.CourseDailyLog;
-import com.m9d.sroom.global.model.CourseVideo;
-import com.m9d.sroom.global.model.Playlist;
-import com.m9d.sroom.global.model.Video;
+import com.m9d.sroom.global.mapper.CourseDailyLog;
+import com.m9d.sroom.global.mapper.CourseVideo;
+import com.m9d.sroom.global.mapper.Playlist;
+import com.m9d.sroom.global.mapper.Video;
 import com.m9d.sroom.lecture.dto.request.KeywordSearchParam;
 import com.m9d.sroom.lecture.dto.request.LectureDetailParam;
 import com.m9d.sroom.lecture.dto.request.LectureTimeRecord;
@@ -32,7 +32,6 @@ import com.m9d.sroom.util.youtube.vo.search.SearchVo;
 import com.m9d.sroom.util.youtube.vo.video.VideoVo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.util.HtmlUtils;
 import reactor.core.publisher.Mono;
