+++ src/main/java/com/m9d/sroom/repository/course/CourseJdbcRepositoryImpl.java
@@ -1,12 +1,9 @@
 package com.m9d.sroom.repository.course;
 
-import com.m9d.sroom.course.dto.response.CourseInfo;
-import com.m9d.sroom.global.model.Course;
-import com.m9d.sroom.lecture.dto.response.CourseBrief;
+import com.m9d.sroom.global.mapper.Course;
 import org.springframework.stereotype.Repository;
 
 import java.util.Date;
-import java.util.List;
 
 @Repository
 public class CourseJdbcRepositoryImpl implements CourseRepository{
