+++ src/main/java/com/m9d/sroom/repository/coursedailylog/CourseDailyLogRepository.java
@@ -1,11 +1,8 @@
 package com.m9d.sroom.repository.coursedailylog;
 
-import com.m9d.sroom.dashboard.dto.response.DashboardMemberData;
-import com.m9d.sroom.dashboard.dto.response.LearningHistory;
-import com.m9d.sroom.global.model.CourseDailyLog;
+import com.m9d.sroom.global.mapper.CourseDailyLog;
 
 import java.sql.Date;
-import java.util.List;
 import java.util.Optional;
 
 public interface CourseDailyLogRepository {
