+++ src/main/java/com/m9d/sroom/repository/coursequiz/CourseQuizRepository.java
@@ -1,11 +1,8 @@
 package com.m9d.sroom.repository.coursequiz;
 
 
-import com.m9d.sroom.global.model.CourseQuiz;
-import com.m9d.sroom.material.model.CourseQuizInfo;
+import com.m9d.sroom.global.mapper.CourseQuiz;
 import com.m9d.sroom.material.model.SubmittedQuizInfo;
-
-import java.util.Optional;
 
 public interface CourseQuizRepository {
