+++ src/main/java/com/m9d/sroom/repository/playlistvideo/PlaylistVideoJdbcRepositoryImpl.java
@@ -1,10 +1,9 @@
 package com.m9d.sroom.repository.playlistvideo;
 
-import com.m9d.sroom.global.model.PlaylistVideo;
+import com.m9d.sroom.global.mapper.PlaylistVideo;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Map;
 
 @Repository
 public class PlaylistVideoJdbcRepositoryImpl implements PlaylistVideoRepository{
