+++ src/main/java/com/m9d/sroom/repository/playlistvideo/PlaylistVideoRepository.java
@@ -1,9 +1,8 @@
 package com.m9d.sroom.repository.playlistvideo;
 
-import com.m9d.sroom.global.model.PlaylistVideo;
+import com.m9d.sroom.global.mapper.PlaylistVideo;
 
 import java.util.List;
-import java.util.Map;
 
 public interface PlaylistVideoRepository {
