+++ src/main/java/com/m9d/sroom/repository/quiz/QuizRepository.java
@@ -1,7 +1,7 @@
 package com.m9d.sroom.repository.quiz;
 
 
-import com.m9d.sroom.global.model.Quiz;
+import com.m9d.sroom.global.mapper.Quiz;
 
 import java.util.List;
