+++ src/main/java/com/m9d/sroom/repository/review/ReviewJdbcRepositoryImpl.java
@@ -1,7 +1,6 @@
 package com.m9d.sroom.repository.review;
 
-import com.m9d.sroom.global.model.Review;
-import com.m9d.sroom.lecture.dto.response.ReviewBrief;
+import com.m9d.sroom.global.mapper.Review;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
