+++ src/main/java/com/m9d/sroom/repository/review/ReviewRepository.java
@@ -1,7 +1,6 @@
 package com.m9d.sroom.repository.review;
 
-import com.m9d.sroom.global.model.Review;
-import com.m9d.sroom.lecture.dto.response.ReviewBrief;
+import com.m9d.sroom.global.mapper.Review;
 
 import java.util.List;
