+++ src/main/java/com/m9d/sroom/review/service/ReviewService.java
@@ -3,7 +3,7 @@ package com.m9d.sroom.review.service;
 import com.m9d.sroom.review.dto.LectureBrief4Review;
 import com.m9d.sroom.review.dto.LectureBriefList4Review;
 import com.m9d.sroom.review.dto.LectureData;
-import com.m9d.sroom.review.dto.Review;
+import com.m9d.sroom.global.mapper.Review;
 import com.m9d.sroom.review.repository.ReviewRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
